# api-server/openapi.yaml
swagger: "2.0"
info:
  description: This is an API to search for employees by a certain criteria and return those that match it.
  version: 1.0.0
  title: WorkForce API
  # put the contact info for your development or API team
  contact:
    email: tainabatistaconde@gmail.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /

paths:
  /employees:
    get:
      summary: get all employees.
      produces:
        - application/json
      responses:
        200:
          description: a list of employees
          schema:
            type: array
            items:
              $ref: "#/definitions/EmployeeItem"
        400:
          description: bad input parameter
    post:
      tags:
        - admins
      summary: adds an inventory item
      operationId: addInventory
      description: Adds an item to the system
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: inventoryItem
          description: Inventory item to add
          schema:
            $ref: "#/definitions/InventoryItem"
      responses:
        201:
          description: item created
        400:
          description: invalid input, object invalid
        409:
          description: an existing item already exists
definitions:
  InventoryItem:
    type: object
    required:
      - id
      - name
      - manufacturer
      - releaseDate
    properties:
      id:
        type: string
        format: uuid
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
      name:
        type: string
        example: Widget Adapter
      releaseDate:
        type: string
        format: date-time
        example: 2016-08-29T09:12:33.001Z
      manufacturer:
        $ref: "#/definitions/Manufacturer"
  Manufacturer:
    required:
      - name
    properties:
      name:
        type: string
        example: ACME Corporation
      homePage:
        type: string
        format: url
        example: https://www.acme-corp.com
      phone:
        type: string
        example: 408-867-5309
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /Taina-conde/example/1.0.0
schemes:
  - https